arm64: Verify PSCI conduit for ARM_SMCCC_ARCH_WORKAROUND_2
In commit f045269b5a597be78d39e9950ff6d1f063bd4681 ("arm64: Call
ARCH_WORKAROUND_2 on transitions between EL0 and EL1"), the co-author who
modified the patch has left out the call to `alternative_cb`, mentioning
the code for it was not present.

The code in question was, in fact, present, but it resulted in a hardcoded
PSCI conduit against the SMC only.

Pull in the rest of the patch by the original author and remove the
hardcoded SMC, taking into account the HVC as well.

Fixes: f045269b5a59 ("arm64: Call ARCH_WORKAROUND_2 on transitions between EL0 and EL1")
Co-authored-by: Marc Zyngier <marc.zyngier@arm.com>
Signed-off-by: voidanix <voidanix@keyedlimepie.org>