arm64: Verify PSCI conduit for ARM_SMCCC_ARCH_WORKAROUND_2
In commit f045269b5a597be78d39e9950ff6d1f063bd4681 ("arm64: Call
ARCH_WORKAROUND_2 on transitions between EL0 and EL1"), the co-author who
modified the patch has left out calls to `alternative_cb`, mentioning the
code for it was not present.

The code in question is, in fact, present, but it nevertheless resulted in
a hardcoded PSCI conduit that used the SMC only.

Pull in the rest of the patch by the original author and use the
`alternative_cb` macro to take the HVC into account as well.

Fixes: f045269b5a59 ("arm64: Call ARCH_WORKAROUND_2 on transitions between EL0 and EL1")
Patch-originally-by: Marc Zyngier <marc.zyngier@arm.com>
Signed-off-by: voidanix <voidanix@keyedlimepie.org>