regulator: Import stub-regulator from K.P.1.0.r1
Also remove its outdated documentation and unused header.

The main point this is to avoid the following WARNING from sanity checks on
vmlinux:

WARNING: vmlinux.o(___ksymtab+regulator_stub_init+0x0): Section mismatch in
reference from the variable __ksymtab_regulator_stub_init to the function
.init.text:regulator_stub_init() The symbol regulator_stub_init is exported
and annotated __init Fix this by removing the __init annotation of
regulator_stub_init or drop the export.

In addition, debug regulator support was also removed as it is not
supported on this kernel.

Signed-off-by: voidanix <voidanix@keyedlimepie.org>