random: keep injecting pre-init bits into the pool
Commit fed6de0e0bac9d446de80b4734820c4b16c2abe9 ("random: do not use
batches when !crng_ready()") leads some of our Sony devices to bootloop
and on top of that, it was a merge between several mainline commits,
making the real issue harder to find.

Kmsg on the device revealed that code inside the previously existing
`account` condition resulted in being unreachable, meaning that our
devices only required an initial injection into the base crng to start
properly.

Revert part of the aforementioned commit, while removing the "account"
parameter and its mentions.

Fixes: fed6de0e0bac ("random: do not use batches when !crng_ready()")
Signed-off-by: voidanix <voidanix@keyedlimepie.org>